refactor : Add updateValue function and Remove form tag
